Move data layer into specified data directory
